requirements.txt
Flask==2.2.3
pytz==2022.7.1
google-cloud-datastore==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-cloud-vision==3.3.1
google-cloud-translate==3.10.1
Pillow
python-dotenv
Flask-Cors
